// WebUI/app-frontend/src/app/enums/training-score-type.ts
export enum TrainingScoreType {
    BallControl = 0,
    LeftFootPassAccuracy = 1,
    RightFootPassAccuracy = 2,
    LeftFootDribblingAccuracy = 3,
    RightFootDribblingAccuracy = 4,
    LeftFootBallReceivingAccuracy = 5,
    RightFootBallReceivingAccuracy = 6,
    LeftFootShotsAccuracy = 7,
    RightFootShotsAccuracy = 8,
    HeadingAccuracy = 9,
    OneVsOneScore = 10,
    Mobility = 11,
    Strength = 12,
    Endurance = 13,
    Agility = 14,
    Coordination = 15,
    Concentration = 16,
    EmotionsControl = 17,
    Selfconfidence = 18,
    StressControl = 19,
    Attitude = 20,
    Communication = 21,
    Cooperation = 22,
    Determination = 23,
    Discipline = 24,
    Engagement = 25,
    Creativity = 26,
    Decisiveness = 27,
    Awareness = 28
}